Event and Tick handlers' names were swapped
